+++ src/test/java/printer/ReceiptPrinterTest.java
@@ -1,3 +1,5 @@
+package printer;
+
 import static org.junit.Assert.assertEquals;
 
 import java.util.LinkedHashMap;
@@ -5,6 +7,10 @@ import java.util.LinkedHashMap;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+
+import model.Product;
+import printer.ReceiptPrinter;
+
 
 public class ReceiptPrinterTest {
